PCHR-3084: Move sass tasks in a separate module
