+++ uk.co.compucorp.civicrm.hrcore/gulpfile.js
@@ -159,73 +159,11 @@ var test = require('./gulp/test');
 
 // Sass
 (function () {
-  var bulk = require('gulp-sass-bulk-import');
-  var civicrmScssRoot = require('civicrm-scssroot')();
-  var sass = require('gulp-sass');
-  var stripCssComments = require('gulp-strip-css-comments');
+  var tasks = require('./gulp/tasks/sass');
 
-  gulp.task('sass', function (cb) {
-    if (hasCurrentExtensionMainSassFile()) {
-      var sequence = utils.addExtensionCustomTasksToSequence([
-        utils.spawnTaskForExtension('sass:sync', syncTask, utils.getCurrentExtension()),
-        utils.spawnTaskForExtension('sass:main', mainTask, utils.getCurrentExtension())
-      ], 'sass');
-
-      gulpSequence.apply(null, sequence)(cb);
-    } else {
-      console.log(colors.yellow('No main .scss file found, skipping...'));
-      cb();
-    }
+  tasks.forEach(function (task) {
+    gulp.task(task.name, task.fn);
   });
-
-  gulp.task('sass:watch', function () {
-    var extPath = utils.getExtensionPath();
-    var watchPatterns = utils.addExtensionCustomWatchPatternsToDefaultList([
-      path.join(extPath, 'scss/**/*.scss')
-    ], 'sass');
-
-    gulp.watch(watchPatterns, ['sass']);
-  });
-
-  /**
-   * Check if the current extension has a main *.scss file
-   *
-   * @return {Boolean}
-   */
-  function hasCurrentExtensionMainSassFile () {
-    return !!find.fileSync(/\/scss\/([^/]+)?\.scss$/, utils.getExtensionPath())[0];
-  }
-
-  /**
-   * Compiles SASS files
-   *
-   * @param {Function} cb
-   * @return {Vinyl}
-   */
-  function mainTask (cb) {
-    var extPath = utils.getExtensionPath();
-
-    return gulp.src(path.join(extPath, '/scss/*.scss'))
-      .pipe(bulk())
-      .pipe(sass({
-        outputStyle: 'compressed',
-        includePaths: civicrmScssRoot.getPath(),
-        precision: 10
-      }).on('error', sass.logError))
-      .pipe(stripCssComments({ preserve: false }))
-      .pipe(gulp.dest(path.join(extPath, '/css/')));
-  }
-
-  /**
-   * Syncs the SASS cache
-   *
-   * @param {Function} cb
-   * @return {Vinyl}*
-   */
-  function syncTask (cb) {
-    civicrmScssRoot.updateSync();
-    cb();
-  }
 }());
 
 // RequireJS
